support GIT getting operation using git bash process (#2611)
* support git exe lib for GIT GET operations

* add validation check

* fix comments

* fix test tmp folder

* fix build error for linux

* rename files

* fix some comments

* change process utility to public
