+++ docfx.sln
@@ -1,15 +1,14 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26124.0
+VisualStudioVersion = 15.0.27130.2003
 MinimumVisualStudioVersion = 15.0.26124.0
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{D9F346F2-6435-4FD4-8980-33233FC585A6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "docfx", "src\docfx\docfx.csproj", "{DDE1E383-87B3-4C05-AF94-5C955035313A}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "docfx", "src\docfx\docfx.csproj", "{DDE1E383-87B3-4C05-AF94-5C955035313A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{5BE2024B-B512-492C-87AB-E8C7D3B3F3AE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "docfx.Test", "test\docfx.Test\docfx.Test.csproj", "{B212F8F3-BE11-45C2-B0EA-097051CBF654}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "docfx.Test", "test\docfx.Test\docfx.Test.csproj", "{B212F8F3-BE11-45C2-B0EA-097051CBF654}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -20,37 +19,40 @@ Global
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Debug|x64.ActiveCfg = Debug|x64
-		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Debug|x64.Build.0 = Debug|x64
-		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Debug|x86.ActiveCfg = Debug|x86
-		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Debug|x86.Build.0 = Debug|x86
+		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Debug|x64.Build.0 = Debug|Any CPU
+		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Debug|x86.Build.0 = Debug|Any CPU
 		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Release|x64.ActiveCfg = Release|x64
-		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Release|x64.Build.0 = Release|x64
-		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Release|x86.ActiveCfg = Release|x86
-		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Release|x86.Build.0 = Release|x86
+		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Release|x64.ActiveCfg = Release|Any CPU
+		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Release|x64.Build.0 = Release|Any CPU
+		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Release|x86.ActiveCfg = Release|Any CPU
+		{DDE1E383-87B3-4C05-AF94-5C955035313A}.Release|x86.Build.0 = Release|Any CPU
 		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Debug|x64.ActiveCfg = Debug|x64
-		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Debug|x64.Build.0 = Debug|x64
-		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Debug|x86.ActiveCfg = Debug|x86
-		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Debug|x86.Build.0 = Debug|x86
+		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Debug|x64.Build.0 = Debug|Any CPU
+		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Debug|x86.Build.0 = Debug|Any CPU
 		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Release|x64.ActiveCfg = Release|x64
-		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Release|x64.Build.0 = Release|x64
-		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Release|x86.ActiveCfg = Release|x86
-		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Release|x86.Build.0 = Release|x86
+		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Release|x64.ActiveCfg = Release|Any CPU
+		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Release|x64.Build.0 = Release|Any CPU
+		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Release|x86.ActiveCfg = Release|Any CPU
+		{B212F8F3-BE11-45C2-B0EA-097051CBF654}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{DDE1E383-87B3-4C05-AF94-5C955035313A} = {D9F346F2-6435-4FD4-8980-33233FC585A6}
 		{B212F8F3-BE11-45C2-B0EA-097051CBF654} = {5BE2024B-B512-492C-87AB-E8C7D3B3F3AE}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {BC590A45-FAD1-4FA6-A5AA-40C7BD265605}
 	EndGlobalSection
 EndGlobal
