+++ src/docfx/lib/git/GitCommit.cs
